requirements.txt
Flask
flask-cors
pandas
openpyxl
matplotlib
seaborn
python-docx
python-dotenv
google-generativeai
flask-bcrypt
flask-sqlalchemy